Feb 21, 2025, 9:47 AM

+++ CT06_06/lesson6.py
@@ -2,8 +2,8 @@
 num1 = input("Give me your start value. ")
 num2 = input("Give me your end value. ")
 
-num1 = float(num1)
-num2 = float(num2)
+num1 = int(num1)
+num2 = int(num2)
 
 if num1 > num2:
     for Russia in range(num2, num1):
